pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Semeikin_Kirill <semeikinkirill21@gmail.com>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
